@@ -2,14 +2,14 @@
   "name": "template",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "index.tsx",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "webpack",
-    "start": "webpack-dev-server --open --mode development"
+    "build": "webpack --mode production --config ./config/webpack.prod.config.js",
+    "start": "webpack-dev-server --open --mode development --config ./config/webpack.dev.config.js"
   },
   "keywords": [],
-  "author": "yogocheng",
+  "author": "wangcheng",
   "license": "ISC",
   "devDependencies": {
     "@babel/core": "^7.8.7",
@@ -18,25 +18,27 @@
     "@babel/preset-env": "^7.8.7",
     "@babel/preset-react": "^7.8.3",
     "@babel/runtime": "^7.8.7",
-    "axios": "^0.19.2",
     "babel-loader": "^8.0.6",
-    "babel-preset-mobx": "^2.0.0",
     "css-loader": "^3.4.2",
-    "html-webpack-plugin": "^3.2.0",
-    "mobx": "^5.15.4",
-    "mobx-react": "^6.1.8",
+    "html-webpack-plugin": "^5.5.0",
     "path": "^0.12.7",
-    "react-router-dom": "^5.1.2",
     "style-loader": "^1.1.3",
     "ts-loader": "^6.2.1",
     "typescript": "^3.8.3",
     "url-loader": "^3.0.0",
-    "webpack": "^4.42.0",
-    "webpack-cli": "^3.3.11",
-    "webpack-dev-server": "^3.10.3"
+    "webpack": "^5.75.0",
+    "webpack-cli": "^4.10.0",
+    "webpack-dev-server": "^4.11.1"
   },
   "dependencies": {
-    "react": "^16.13.0",
-    "react-dom": "^16.13.0"
+    "@types/react": "^18.0.25",
+    "@types/react-dom": "^18.0.9",
+    "clean-webpack-plugin": "^4.0.0",
+    "mini-css-extract-plugin": "^2.6.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.4.3",
+    "sass": "^1.56.1",
+    "sass-loader": "^13.2.0"
   }
 }
